requirements.txt
asgiref==3.4.1
attrs==21.2.0
Django==3.2.7
django-webpack-loader==1.4.0
djangorestframework==3.12.4
drf-spectacular==0.19.0
inflection==0.5.1
jsonschema==3.2.0
mypy==0.910
mypy-extensions==0.4.3
nodeenv==1.6.0
pyrsistent==0.18.0
pytz==2021.1
PyYAML==5.4.1
six==1.16.0
sqlparse==0.4.2
toml==0.10.2
typing-extensions==3.10.0.2
uritemplate==3.0.1
